moved the get_drive_ind from the raw module to the data module.
